requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
altair==5.0.1
anyio==3.7.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
azure-ai-formrecognizer==3.3.0b1
azure-common==1.1.28
azure-core==1.27.1
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
boilerpy3==1.0.6
bs4==0.0.1
canals==0.2.2
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
databricks-cli==0.17.7
datasets==2.12.0
dill==0.3.6
docker==6.1.3
docopt==0.6.2
elasticsearch==7.17.9
entrypoints==0.4
evaluate==0.4.0
Events==0.4
farm-haystack==1.17.1
fastapi==0.98.0
ffmpy==0.3.0
filelock==3.12.2
Flask==2.3.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
generalimport==0.3.1
gitdb==4.0.10
GitPython==3.1.31
gradio==3.35.2
gradio_client==0.2.7
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
inflect==6.0.4
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langdetect==1.0.9
linkify-it-py==2.0.2
lit==16.0.5.post0
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mlflow==2.4.1
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
num2words==0.5.12
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
orjson==3.9.1
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
posthog==3.0.1
protobuf==3.20.2
psutil==5.9.5
pyarrow==12.0.1
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
quantulum3==0.9.0
querystring-parser==1.2.4
rank-bm25==0.2.2
rapidfuzz==2.7.0
regex==2023.6.3
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
responses==0.18.0
rouge-score==0.1.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
sqlparse==0.4.4
sseclient-py==1.7.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
url-normalize==1.4.3
urllib3==1.26.16
uvicorn==0.22.0
websocket-client==1.5.3
websockets==11.0.3
Werkzeug==2.3.6
xformers==0.0.20
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
